requirements.txt
slack_bolt
python-dotenv
redis
requests